Redesign item form
Redesign the item form so that it looks better. Extra spacing is addded
to make it more visually appealing. Also the save button is now themed
as well as bigger to make it easier for users to see and use it.
Additionally, the date field for expiry is removed since it doesn't
make sense to put the expiry when you are adding items. It should be
relocated to the fridge section.

see #67